+++ src/EnigmeTuyaux.h
@@ -10,8 +10,8 @@ class EnigmeTuyaux : public Enigme
 {
 public:
 	explicit EnigmeTuyaux(int doorId);
-	void addTuyauFixe(std::string const& type,int orientation, std::pair<int,int> const& position);
-	void addTuyauMobile(int id,std::string const& type, int orientation, std::pair<int, int> const& position, std::vector<int> const& switchs);
+	void addTuyauFixe(std::string const& type,int orientation, std::pair<float,float> const& position);
+	void addTuyauMobile(int id,std::string const& type, int orientation, std::pair<float,float> const& position, std::vector<int> const& switchs);
 	void addTuyauSolution(int id, std::vector<int>const& solutions);
 
 	void draw(sf::RenderWindow& window) override;
